correcting var_decl to be explicitly defined int32

+++ +helics/HELICS_FALSE.m
@@ -4,7 +4,7 @@ indicator used for a false response
 function v = HELICS_FALSE()
 	persistent vInitialized;
 	if isempty(vInitialized)
-		vInitialized = 0;
+		vInitialized = int32(0);
 	end
 	v = vInitialized;
 end
